# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AssetMind Backend API",
        "description": "REST API untuk Sistem Pendukung Keputusan Perawatan Aset Industri Perkebunan Kelapa Sawit. Backend Laravel terintegrasi dengan AI Engine (FastAPI) untuk embedding, chatbot RAG, dan rekomendasi pemeliharaan.",
        "contact": {
            "name": "AssetMind",
            "email": "firasyan.daffa123@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Server lokal (Laravel artisan serve)"
        }
    ],
    "paths": {
...
        "/api/repair-requests": {
            "get": {
                "tags": [
                    "Repair Request"
                ],
                "summary": "Daftar pengajuan perbaikan",
                "description": "Mekanik hanya melihat pengajuan miliknya sendiri.",
                "operationId": "repairRequestIndex",
                "responses": {
                    "200": {
                        "description": "Data pengajuan perbaikan berhasil diambil"
                    }
                },
                "security": [
                    {
                        "bearerSanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Repair Request"
                ],
                "summary": "Buat pengajuan perbaikan (khusus mekanik)",
                "operationId": "repairRequestStore",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "asset_id",
                                    "kondisi_perbaikan",
                                    "catatan_kerusakan"
                                ],
                                "properties": {
                                    "asset_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "kondisi_perbaikan": {
                                        "type": "string",
                                        "example": "sedang",
                                        "enum": [
                                            "ringan",
                                            "sedang",
                                            "berat"
                                        ]
                                    },
                                    "catatan_kerusakan": {
                                        "type": "string",
                                        "example": "Seal bocor pada silinder"
                                    },
                                    "bukti_foto": {
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pengajuan perbaikan berhasil dibuat"
                    },
                    "422": {
                        "description": "Validasi gagal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerSanctum": []
                    }
                ]
            }
        },
        "/api/repair-requests/{id}": {
            "get": {
                "tags": [
                    "Repair Request"
                ],
                "summary": "Detail pengajuan perbaikan",
                "operationId": "repairRequestShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail pengajuan perbaikan berhasil diambil"
                    },
                    "403": {
                        "description": "Bukan milik Anda",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerSanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Repair Request"
                ],
                "summary": "Perbarui pengajuan perbaikan (hanya saat status pending)",
                "operationId": "repairRequestUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "asset_id",
                                    "kondisi_perbaikan",
                                    "catatan_kerusakan"
                                ],
                                "properties": {
                                    "asset_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "kondisi_perbaikan": {
                                        "type": "string",
                                        "example": "berat",
                                        "enum": [
                                            "ringan",
                                            "sedang",
                                            "berat"
                                        ]
                                    },
                                    "catatan_kerusakan": {
                                        "type": "string"
                                    },
                                    "bukti_foto": {
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pengajuan perbaikan berhasil diperbarui"
                    },
                    "403": {
                        "description": "Bukan milik Anda",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Sudah diverifikasi / validasi gagal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerSanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Repair Request"
                ],
                "summary": "Hapus pengajuan perbaikan (hanya saat status pending)",
                "operationId": "repairRequestDestroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pengajuan perbaikan berhasil dihapus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Bukan milik Anda",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Sudah diverifikasi"
                    }
                },
                "security": [
                    {
                        "bearerSanctum": []
                    }
                ]
            }
        },
        "/api/repair-requests/{id}/verify": {
            "patch": {
                "tags": [
                    "Repair Request"
                ],
                "summary": "Verifikasi pengajuan perbaikan (khusus teknik)",
                "description": "Hanya pengajuan berstatus pending yang bisa diverifikasi menjadi approved/rejected.",
                "operationId": "repairRequestVerify",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status_verifikasi"
                                ],
                                "properties": {
                                    "status_verifikasi": {
                                        "type": "string",
                                        "example": "approved",
                                        "enum": [
                                            "approved",
                                            "rejected"
                                        ]
                                    },
                                    "catatan_verifikasi": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status pengajuan perbaikan berhasil diverifikasi"
                    },
                    "422": {
                        "description": "Sudah diverifikasi / validasi gagal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerSanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "MessageResponse": {
                "properties": {
                    "message": {
                        "description": "Definisi dasar dokumentasi OpenAPI/Swagger untuk AssetMind Backend API.\n\nMemakai PHP 8 Attributes (didukung swagger-php v6).\nAnotasi path tersebar di controller masing-masing.",
                        "type": "string",
                        "example": "Operasi berhasil"
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "ForbiddenError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Forbidden. Data ini bukan milik Anda."
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerSanctum": {
                "type": "http",
                "description": "Masukkan token Sanctum dari endpoint /auth/login (tanpa kata 'Bearer').",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}